Conflicts:
	game/src/game/Engine.java
	game/src/gui/BoardPanel.java
	game/src/gui/GUI.java
	game/src/gui/HighScoresGUI.java
	game/src/gui/MenuGUI.java
	game/src/gui/PlayGUI.java
	game/src/gui/SettingsGUI.java

+++ game/src/gui/HighScoresGUI.java
@@ -75,95 +75,105 @@ public class HighScoresGUI extends JFrame {
 		switch (highScoresObject.getPlayerListSize()) {
 		case 1:
 			labels[0][0].setText(highScoresObject.getPlayer(1).getName());
-			labels[0][1].setText((Integer.toString(highScoresObject
+			labels[0][1].setText((Double.toString(highScoresObject
 					.getPlayer(1).getScore())));
 			labels[0][2].setText(dateFormat.format(highScoresObject
 					.getPlayer(1).getDate()));
 			break;
 		case 2:
 			labels[0][0].setText(highScoresObject.getPlayer(1).getName());
-			labels[0][1].setText((Integer.toString(highScoresObject
+			labels[0][1].setText((Double.toString(highScoresObject
 					.getPlayer(1).getScore())));
 			labels[0][2].setText(dateFormat.format(highScoresObject
 					.getPlayer(1).getDate()));
 
 			labels[1][0].setText(highScoresObject.getPlayer(2).getName());
-			labels[1][1].setText((Integer.toString(highScoresObject
+
+			labels[1][1].setText((Double.toString(highScoresObject
 					.getPlayer(2).getScore())));
 			labels[1][2].setText(dateFormat.format(highScoresObject
 					.getPlayer(2).getDate()));
 			break;
 		case 3:
 			labels[0][0].setText(highScoresObject.getPlayer(1).getName());
-			labels[0][1].setText((Integer.toString(highScoresObject
+			labels[0][1].setText((Double.toString(highScoresObject
 					.getPlayer(1).getScore())));
 			labels[0][2].setText(dateFormat.format(highScoresObject
 					.getPlayer(1).getDate()));
 
 			labels[1][0].setText(highScoresObject.getPlayer(2).getName());
-			labels[1][1].setText((Integer.toString(highScoresObject
+
+			labels[1][1].setText((Double.toString(highScoresObject
 					.getPlayer(2).getScore())));
 			labels[1][2].setText(dateFormat.format(highScoresObject
 					.getPlayer(2).getDate()));
 
 			labels[2][0].setText(highScoresObject.getPlayer(3).getName());
-			labels[2][1].setText((Integer.toString(highScoresObject
+			labels[2][1].setText((Double.toString(highScoresObject
 					.getPlayer(3).getScore())));
 			labels[2][2].setText(dateFormat.format(highScoresObject
 					.getPlayer(3).getDate()));
 			break;
 		case 4:
 			labels[0][0].setText(highScoresObject.getPlayer(1).getName());
-			labels[0][1].setText((Integer.toString(highScoresObject
+			labels[0][1].setText((Double.toString(highScoresObject
 					.getPlayer(1).getScore())));
 			labels[0][2].setText(dateFormat.format(highScoresObject
 					.getPlayer(1).getDate()));
 
 			labels[1][0].setText(highScoresObject.getPlayer(2).getName());
-			labels[1][1].setText((Integer.toString(highScoresObject
+
+			labels[1][1].setText((Double.toString(highScoresObject
 					.getPlayer(2).getScore())));
 			labels[1][2].setText(dateFormat.format(highScoresObject
 					.getPlayer(2).getDate()));
 
 			labels[2][0].setText(highScoresObject.getPlayer(3).getName());
-			labels[2][1].setText((Integer.toString(highScoresObject
+			labels[2][1].setText((Double.toString(highScoresObject
+
 					.getPlayer(3).getScore())));
 			labels[2][2].setText(dateFormat.format(highScoresObject
 					.getPlayer(3).getDate()));
 
 			labels[3][0].setText(highScoresObject.getPlayer(4).getName());
-			labels[3][1].setText((Integer.toString(highScoresObject
+
+			labels[3][1].setText((Double.toString(highScoresObject
+
 					.getPlayer(4).getScore())));
 			labels[3][2].setText(dateFormat.format(highScoresObject
 					.getPlayer(4).getDate()));
 			break;
 		case 5:
 			labels[0][0].setText(highScoresObject.getPlayer(1).getName());
-			labels[0][1].setText((Integer.toString(highScoresObject
+			labels[0][1].setText((Double.toString(highScoresObject
 					.getPlayer(1).getScore())));
 			labels[0][2].setText(dateFormat.format(highScoresObject
 					.getPlayer(1).getDate()));
 
 			labels[1][0].setText(highScoresObject.getPlayer(2).getName());
-			labels[1][1].setText((Integer.toString(highScoresObject
+			labels[1][1].setText((Double.toString(highScoresObject
 					.getPlayer(2).getScore())));
 			labels[1][2].setText(dateFormat.format(highScoresObject
 					.getPlayer(2).getDate()));
 
 			labels[2][0].setText(highScoresObject.getPlayer(3).getName());
-			labels[2][1].setText((Integer.toString(highScoresObject
+			labels[2][1].setText((Double.toString(highScoresObject
+
 					.getPlayer(3).getScore())));
 			labels[2][2].setText(dateFormat.format(highScoresObject
 					.getPlayer(3).getDate()));
 
 			labels[3][0].setText(highScoresObject.getPlayer(4).getName());
-			labels[3][1].setText((Integer.toString(highScoresObject
+
+			labels[3][1].setText((Double.toString(highScoresObject
+
 					.getPlayer(4).getScore())));
 			labels[3][2].setText(dateFormat.format(highScoresObject
 					.getPlayer(4).getDate()));
 
 			labels[4][0].setText(highScoresObject.getPlayer(5).getName());
-			labels[4][1].setText((Integer.toString(highScoresObject
+
+			labels[4][1].setText((Double.toString(highScoresObject
 					.getPlayer(5).getScore())));
 			labels[4][2].setText(dateFormat.format(highScoresObject
 					.getPlayer(5).getDate()));
